+++ src/app/shared/services/local-storage.service.ts
@@ -1,0 +1,21 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LocalStorageService {
+
+  constructor() { }
+
+  save(key: string, value: string) {
+    localStorage.setItem(key, value);
+  }
+
+  get(key: string): string {
+    return localStorage.getItem(key);
+  }
+
+  remove(key: string) {
+    localStorage.removeItem(key);
+  }
+}