+++ changelog.md
@@ -29,3 +29,4 @@ Task 3
 ----------------------------------
 ###Added
 * Order service
+* Local storage service